+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @did-btcr2/cli
 
+## 0.23.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @did-btcr2/method@0.61.0
+  - @did-btcr2/api@0.23.1
+
 ## 0.23.1
 
 ### Patch Changes